+++ study/api-src/org/labkey/api/specimen/actions/IdTypes.java
@@ -1,0 +1,8 @@
+package org.labkey.api.specimen.actions;
+
+public enum IdTypes
+{
+    GlobalUniqueId,
+    SpecimenHash,
+    RowId
+}
