+++ study/src/org/labkey/study/view/specimen/webPart.jsp
@@ -19,7 +19,6 @@
 <%@ page import="org.labkey.api.security.User"%>
 <%@ page import="org.labkey.api.security.permissions.AdminPermission"%>
 <%@ page import="org.labkey.api.specimen.SpecimenMigrationService"%>
-<%@ page import="org.labkey.api.specimen.actions.ShowSearchAction" %>
 <%@ page import="org.labkey.api.specimen.security.permissions.RequestSpecimensPermission" %>
 <%@ page import="org.labkey.api.specimen.settings.SettingsManager" %>
 <%@ page import="org.labkey.api.specimen.view.SpecimenWebPart" %>
@@ -181,7 +180,7 @@
                             <tbody>
                             <tr class="labkey-nav-tree-row labkey-header">
                                 <td class="labkey-nav-tree-text">
-                                    <a href="<%=h(new ActionURL(ShowSearchAction.class,c).addParameter("showVials","false"))%>">For Vial Groups</a></td>
+                                    <a href="<%=h(SpecimenMigrationService.get().getShowSearchURL(getContainer(), false))%>">For Vial Groups</a></td>
                             </tr>
                             </tbody>
                         </table>
@@ -193,7 +192,7 @@
                             <tbody>
                             <tr class="labkey-nav-tree-row labkey-header">
                                 <td class="labkey-nav-tree-text">
-                                    <a href="<%=h(new ActionURL(ShowSearchAction.class,c).addParameter("showVials","true"))%>">For Individual Vials</a></td>
+                                    <a href="<%=h(SpecimenMigrationService.get().getShowSearchURL(getContainer(), true))%>">For Individual Vials</a></td>
                             </tr>
                             </tbody>
                         </table>
